requirements.txt
absl-py==1.2.0
beautifulsoup4==4.11.1
cachetools==5.2.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
cycler==0.11.0
Cython==0.29.32
filelock==3.8.0
flake8==5.0.4
Flask==2.2.2
Flask-Cors==3.0.10
fonttools==4.36.0
future==0.18.2
gdown==4.5.1
google-auth==2.10.0
google-auth-oauthlib==0.4.6
grpcio==1.47.0
h5py==3.7.0
idna==3.3
imageio==2.21.1
importlib-metadata==4.12.0
isort==4.3.21
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.5.3
mccabe==0.7.0
numpy==1.23.2
oauthlib==3.2.0
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.9.1
pyflakes==2.5.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.9.0
six==1.16.0
soupsieve==2.3.2.post1
tb-nightly==2.11.0a20220819
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
torch==1.12.1
-e git+ssh://git@github.com/BimanCSE/PersonReidentification.git@646c642b296130729c6bdec5dc7452844454dc09#egg=torchreid&subdirectory=deep-person-reid
torchvision==0.13.1
tqdm==4.64.0
typing_extensions==4.3.0
urllib3==1.26.11
uWSGI==2.0.20
Werkzeug==2.2.2
yacs==0.1.8
yapf==0.32.0
zipp==3.8.1
